Version 0.1.6.0 (#23)
* Save the last browser tabs and sessions.
Add BrowserSettings and BrowserSettingsHelper to save and load the settings and save settings in JSON.
Add Bookmark and TabSettingsModel as models for settings.
Update version to 0.1.4.
Call Cef.Shutdown(); on application exit event.

* Update README.md

* Update README.md

Update image links.

* Improve documentation.
Increase version in documentation and installer.
Move all CefSharp handlers in Handlers folder.
Prepare to download.

* Add Title in setting that is used in the initial tab browser header on loading the tab settings.
Improve loading and saving the settings.
Save the selected browser tab index and loading on loading.
Fix cannot close URL editor on mouse click outside the editor.

* Preparing version 0.1.4.
Fix problem cannot collapse Video quality by adding IsTexBoxFocused property.
Prevent exeption if the JSON setting has no data by checkng whether tab setting has any item.
Improve documentation and adding feature list.
Update setup by incresing folder version to 0.1.4.

* Add DownloadItemModel as base class for download video and classic download items.
Replace ProcessModel to DownloadItemModel as base class for list.
Update handler namespaces.
Add ApplicationIconHelper for getting image from file extension.
Remove disable-gpu-compositing in Cef settings to improve the performance.
Prepare in supporting classing download.
Move shared code from OperationModel to its new base class DownloadItemModel.

* Downloading is already working properly.
Delete GlobalData and move the rest code to GlobalBrowserData.
Remove DownloadFolder settings.

* Move CefSharpBrowser related code to CefSharpBrowser folder

* Add associated application icon in the download item.
Add StringToImageConverter to convert file extension to ImageSource.
Add ProcessHelper for running the external application to open the desired files.
Adjust the layout.
Update Gemfile.lock.

* Save the completed download items and reload on starting of the application.
Adjust namespaces.

* Update to version 0.1.5.
Sign the assemblies and replace Dragablz with the signed assembly.
Add signature keys.
Update Nugate test packages.
Update installer file version.
Add downloaded item at the beginning from the list.
Remove unused html and js for CefSharpTest that are not used.
Run process in a new Task.
Update documentation.
Fix unit test.

* Improvekey signature to enable unit test in Azure Pipeline.

* Downgrade test nuget to make the DevOps not fail.

* Minor update.

* Update unit test framework again to 2.1.2.

* Downgrade .Net Framework test to 4.6.1.

* Update CefSharp 83.4.20.
Update Test and Test app to .net framework 4.8.
Delete unused key.pfx.

* Update v0.1.5 documentation.
Minor bug fixing by adjusting the folder.

* Minor installer update.

* Initial commit for adding bookmark and improve url text box.
Add BookmarkAddin and IsSecureAddIn for additional functionality in TextBoxUrl.
Add BrowserIcons for all icons in the browser.
Add UrlTextBox to improve the TextBox by adding additional functionality like IsSecure and Bookmarking.

* Version 0.1.6.0
Update youtube-dl and ffmpeg.exe..
Fix exception on AudioBitRate conversion, it was changed to double.
Update CefSharp.Wpf to 94.4.20 and adjust the code.
Update Nuget references.

+++ src/VideoBrowser/VideoBrowser/Core/VideoFormat.cs
@@ -55,7 +55,7 @@
         /// <summary>
         /// Gets the audio bit rate. Returns -1 if not defined.
         /// </summary>
-        public int AudioBitRate { get; private set; }
+        public double AudioBitRate { get; private set; }
 
         /// <summary>
         /// Gets a value indicating whether AudioOnly
@@ -258,7 +258,7 @@
             var abr = token.SelectToken("abr");
             if (abr != null)
             {
-                this.AudioBitRate = int.Parse(abr.ToString());
+                this.AudioBitRate = double.Parse(abr.ToString());
             }
 
             // Check for filesize token
